+++ api/src/main/java/org/allaymc/api/block/interfaces/BlockCopperChainBehavior.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.block.interfaces;
 
 import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.component.BlockOxidationComponent;
 
-public interface BlockCopperChainBehavior extends BlockBehavior {
+public interface BlockCopperChainBehavior extends BlockBehavior, BlockOxidationComponent {
 }