+++ server/src/main/java/org/allaymc/server/block/impl/BlockCopperChainBehaviorImpl.java
@@ -1,5 +1,7 @@
 package org.allaymc.server.block.impl;
 
+import lombok.experimental.Delegate;
+import org.allaymc.api.block.component.BlockOxidationComponent;
 import org.allaymc.api.block.interfaces.BlockCopperChainBehavior;
 import org.allaymc.api.component.Component;
 import org.allaymc.server.component.ComponentProvider;
@@ -7,6 +9,9 @@ import org.allaymc.server.component.ComponentProvider;
 import java.util.List;
 
 public class BlockCopperChainBehaviorImpl extends BlockBehaviorImpl implements BlockCopperChainBehavior {
+    @Delegate
+    protected BlockOxidationComponent oxidationComponent;
+
     public BlockCopperChainBehaviorImpl(
             List<ComponentProvider<? extends Component>> componentProviders) {
         super(componentProviders);