+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Implemented lightning rods.
 - (API) Implemented (copper) lanterns.
 - (API) Implemented copper golem statues.
+- (API) Implemented copper chains.
 
 ### Changed